Numerous updates to make plotting better, read full log
We now have the concept of plot overflow. This is used to ensure we graph
everything the user wants to see by plotting extra off the left and right ends
of the x interval.

The plot_f* functions now take both an X and Y interval. Because of this, we
have introduced a function which will return an appropriate Y interval given a
function and an X interval.

The Y_X bits have been reverted to YofX because the function pointer typedef
kept the "of" form and I think it looks better again :)